+++ bin/KGbuilder/KGbuilder_triplesMaker.js
@@ -132,6 +132,8 @@ var KGbuilder_triplesMaker = {
                                                     p: propertyStr,
                                                     o: objectStr
                                                 });
+                                            }else{
+                                                var x=3
                                             }
                                         } /* else{
                     console.log( "missing " +subjectStr +" "+ propertyStr +" "+  objectStr)
@@ -159,7 +161,7 @@ var KGbuilder_triplesMaker = {
     getTripleSubject: function (tableMappings, mapping, line, callback) {
         //get value for Subject
         var subjectStr = null;
-        if (mapping.subjectIsSpecificUri) {
+        if (mapping.subjectIsSpecificUri || mapping.s.endsWith("_#")) {
             subjectStr = mapping.s;
         } else if (typeof mapping.s === "function") {
             try {
@@ -249,7 +251,7 @@ var KGbuilder_triplesMaker = {
         if (mapping.o === "_rowIndex") {
             objectStr = KGbuilder_triplesMaker.getBlankNodeId("_rowIndex");
             return objectStr;
-        } else if (mapping.objectIsSpecificUri) {
+        } else if (mapping.objectIsSpecificUri || mapping.o.endsWith("_#")) {
             objectStr = mapping.o;
         }
