--- public/vocables/modules/tools/KGconstraints/html/KGconstraintsModeler_LeftPanel.html
@@ -1,32 +1,0 @@
-<div style="width: 100%; overflow: auto">
-    <div style="display: flex; flex-direction: row">
-        <div id="KGconstraintsModeler_mainDiv" style="display: none"></div>
-        <div style="display: flex; flex-direction: row" id="KGconstraintsModeler_newAxiomPanel">
-            <!--  <button onclick="KGconstraintsModeler.actions.setLayout('randomlayout')">Random Layout</button>-->
-            <div id="KGconstraintsModeler_activeLegendDiv" style="width: 150px; height: 300px; background-color: #ddd; margin-top: 100px"></div>
-            <div>
-                <div class="input-group-prepend">
-                    <input onkeyup="JstreeWidget.filterTree($(this),'suggestionsSelectJstreeDiv')" class="searchButton" />
-                </div>
-                <div id="suggestionsSelectJstreeDiv" style="width: 300px; height: 500px; overflow: auto"></div>
-            </div>
-        </div>
-    </div>
-    <div>
-        <button class="w3-button slsv-button-3" onclick="Cfihos_pump_poc.query()">Test</button>
-        <button class="w3-button slsv-button-3" onclick="KGconstraints_validator.test()">transform</button>
-
-        <button class="w3-button slsv-button-3" onclick="KGconstraintsModeler.clearConstraints()">clear</button>
-        <!--  <button class="w3-button slsv-button-3" onclick="KGconstraintsModeler_activeLegend.showTriples()">triples</button>
-        <button class="w3-button slsv-button-3" onclick="KGconstraintsModeler.classesDialog()">mappingsTechnicalView</button>
-        <button class="w3-button slsv-button-3" onclick="KGconstraintsModeler.generateBasicMappings()">generateMappings</button>-->
-
-        <button class="w3-button slsv-button-3" onclick="KGconstraintsModeler.saveVisjsGraph()">save graph</button>
-        <button class="w3-button slsv-button-3" onclick="KGconstraintsModeler.loadVisjsGraph()">load graph</button>
-    </div>
-    <div>
-        <button onclick="Cfihos_pump_poc.listEquipments()">List Equipments</button>
-        <input onkeyup="JstreeWidget.filterTree($(this),'KGconstraintsModeler_equipmentsJstreeDiv')" class="searchButton" />
-        <div id="KGconstraintsModeler_equipmentsJstreeDiv" style="width: 350px; height: 500px; overflow: auto"></div>
-    </div>
-</div>
