--- public/vocables/modules/tools/KGconstraints/html/KGconstraintsModeler_graphDiv.html
@@ -1,7 +1,0 @@
-<style></style>
-
-<div id="KGconstraintsModeler_graphDiv" style="height: 95%; background-color: #ddd; margin: 5px; padding: 5px; overflow: auto"></div>
-
-<div id="KGconstraintsModeler_manchesterDiv" style="display: none">
-    <span id="KGconstraintsModeler_textDiv" style="width: 80%; height: 50px; background-color: #ddd"></span>
-</div>
