--- public/vocables/modules/tools/KGconstraints/html/centralPanel.html
@@ -1,10 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-    <head>
-        <meta charset="UTF-8" />
-        <title>Title</title>
-    </head>
-    <body>
-        xxx
-    </body>
-</html>
