--- public/vocables/modules/tools/KGconstraints/html/leftPanel.html
@@ -1,4 +1,0 @@
-<div>
-    <button onclick="KGconstraints_editor.startConstraintEditorBot()">new constraint</button>
-    <div id="KGconstraintsEditor_constraintsTreeDiv" style="width: 350px; height: 500px; overflow: auto"></div>
-</div>
