+++ public/vocables/modules/tools/mappingModeler/html/detailsDialog.html
