+++ public/vocables/modules/tools/mappingModeler/html/mappingModeler_graphDiv.html
@@ -6,7 +6,7 @@
     <button class="w3-button slsv-button-3" onclick="MappingModeler.saveVisjsGraph()">Save </button>
     <button class="w3-button slsv-button-3" onclick="MappingModeler.clearMappings()">Clear</button>
     <button class="w3-button slsv-button-3" onclick="MappingModeler.loadVisjsGraph()">ReLoad</button>
-    <button class="w3-button slsv-button-3" onclick="MappingModeler.classesDialog()"> Details </button>
+    <button class="w3-button slsv-button-3" onclick="MappingsDetails.showDetailsDialog()"> Details </button>
